Move captain nameversion type to own pkg
