+++ internal/runners/platforms/platforms.go
@@ -7,14 +7,14 @@ import (
 
 	"github.com/go-openapi/strfmt"
 
-	"github.com/ActiveState/cli/internal/captain"
+	"github.com/ActiveState/cli/internal/joint"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/output"
 	"github.com/ActiveState/cli/pkg/platform/model"
 )
 
 type PlatformVersion struct {
-	captain.NameVersion
+	joint.NameVersion
 }
 
 func (pv *PlatformVersion) Set(arg string) error {
